BUG/ProductService - 주문 생성시 상품조회 에러 해결, 카프카 메세지 구독 에러 해결, 테스트시 레디스 테스트 제외 (#318)

+++ product-service/src/main/java/com/palja/product_service/infrastructure/repository/DslProductRepository.java
@@ -29,7 +29,7 @@ public class DslProductRepository {
                         product.id,
                         product.companyUserId,
                         product.price.amount.longValue(),
-                        product.productStock.quantity.longValue()))
+                        product.productStock.quantity))
                 .from(product)
                 .where(product.id.eq(productId).and(product.deletedAt.isNull()))
                 .fetchOne();
